* update rubocop to exclude vendor directory * split group notifications to separate module * update application_mailer.rb to conform to rubocop
