+++ app/helpers/application_mailer_helper.rb
@@ -65,4 +65,30 @@ module ApplicationMailerHelper
   def you?(recipient, data)
     recipient.name == data['user']
   end
+
+  def commented_model(data)
+    data['type'].match(/comment_on_([^_]+)/)[1]
+  end
+
+  def extract_comment_notify_key(data)
+    # TODO: all these methods in ApplicationMailerHelper use data['type'],
+    # so just pass the type
+    if comment_on_moment_private(data) || comment_on_strategy_private(data)
+      val('comment_on_private_body')
+    else
+      val('comment_on_body')
+    end
+  end
+
+  def extract_comment_notify_subject(data)
+    # TODO: all these methods in ApplicationMailerHelper use data['type'],
+    # so just pass the type
+    if comment_on_moment(data) || comment_on_moment_private(data)
+      comment_on_moment_subject(data)
+    elsif comment_on_strategy(data) || comment_on_strategy_private(data)
+      comment_on_strategy_subject(data)
+    else
+      comment_on_meeting_subject(data)
+    end
+  end
 end
